some helpful comments and gitignore updates

+++ src/backend/matrixhandler.cpp
@@ -2,6 +2,8 @@
 #include <fstream>
 #include <cstdint>
 
+// Thing the app produces the textfile for the operations
+
 
 // float32_t** createMatrix(int rows, int cols) {
 //     float32_t** matrix = new float32_t*[rows];
@@ -12,7 +14,7 @@
 // }
 
 
-void writetofile(uint8_t size) {
+void writetofile(int size) {
     std::ofstream output;
     std::string nowsize = std::to_string(size);
     output.open("output.txt");
